+++ lib/core.dart
@@ -7,6 +7,7 @@ export 'src/core/len.dart';
 export 'src/core/num.dart';
 export 'src/core/pair.dart';
 export 'src/core/req.dart';
+export 'src/core/req_digit.dart';
 export 'src/core/req_len.dart';
 export 'src/core/rules.dart';
 export 'src/core/to_val_obj.dart';